JVM init arguments type.
The type uses JNI global JNI_GetDefaultJavaVMInitArgs function and so requires to link to libjvm. Because of this, this PR adds a build script for finding it.

Mocking and verifying calls to global functions is not easy in Rust. To do that with JNI functions this PR includes a `raw` module that is supposed to proxy global function calls to the `jni_sys` package in prod and provide a way to mock them in tests.