Change semantics of the Exception token.
Now if the `Exception` token is present there MUST be a pending exception. This makes the code much simpler and I don't think there is actually a case when this is not true. But even if there is, we can always check for the exception manually.

This PR also introduces the `JavaResult` type, reimplements `Class::find` with it and adds an integration test that searches for different classes.